django/src/requirements.txt
django==1.8.18
django-modeltranslation
django-auditlog
django-modeladmin-reorder
django-logentry-admin
django-tinymce
django-user-agents
django-localeurl
django_compressor
django-admin-honeypot
django-mobile
django-defender
django-grappelli==2.10.1
django-filebrowser
django-appconf
django-ratelimit
django-session-security
django-debug-toolbar
django-haystack
django-hitcount
django-htmlmin
django-model-utils
django-simple-history
whoosh
pytz
Pillow
rjsmin
pytils
requests
psycopg2
uwsgi
py-dateutil
rq
django-extensions
django-rq
